refactor(config): move validator tests into dedicated module
